+++ pyextdirect/configuration.py
@@ -17,7 +17,7 @@
 # along with pyextdirect.  If not, see <http://www.gnu.org/licenses/>.
 
 
-__all__ = ['BASIC', 'LOAD', 'SUBMIT', 'STORE_READ', 'STORE_DESTROY', 'ConfigurationMeta', 'create_configuration', 'expose']
+__all__ = ['BASIC', 'LOAD', 'SUBMIT', 'STORE_READ', 'STORE_CUD', 'ConfigurationMeta', 'create_configuration', 'expose']
 
 #: Basic method
 BASIC = 0
@@ -31,8 +31,8 @@ SUBMIT = 2
 #: DirectStore read method
 STORE_READ = 3
 
-#: DirectStore destroy method
-STORE_DESTROY = 4
+#: DirectStore create-update-destroy methods
+STORE_CUD = 4
 
 
 class ConfigurationMeta(type):
@@ -132,3 +132,4 @@ def merge_configurations(configurations):
                 raise ValueError('%s already in a previous base configuration' % k)
             configuration[k] = v
     return configuration
+