+++ pyextdirect/infos.py
@@ -15,4 +15,5 @@
 #
 # You should have received a copy of the GNU Lesser General Public License
 # along with pyextdirect.  If not, see <http://www.gnu.org/licenses/>.
-__version__ = '0.2.2'
+__version__ = '0.2.3'
+