+++ NEWS.rst
@@ -1,3 +1,13 @@
+0.2.3
+=====
+
+* Rename DirectStore destroy to DirectStore CUD for create-update-destroy operations
+
+0.2.2
+=====
+
+* Fix DirectStore destroy
+
 0.2.1
 =====